Added fix for memory leak
Added new target for lab machines

Small fix for 2D arrays

Fixed vector mem leaks

Replaced isLVal function

update runtime function name

refactor: Update backend stack management, fix tuple index validation, and adjust CI configurations.

refactor: Uniquely rename runtime helper functions and remove the dummy print function.

fix: prevent memory leaks in array and vector assignments and refactor stack element popping logic.

fix: free memory from lval during assignment

+++ src/ast/walkers/ValidationWalker.cpp
@@ -329,7 +329,7 @@ ValidationWalker::visitTupleElementAssign(std::shared_ptr<statements::TupleEleme
   if (not tupleTypeSymbol)
     throw TypeError(ctx->getLineNumber(), "Identifier symbol is not a tuple type");
   if (ctx->getFieldIndex() == 0 ||
-      ctx->getFieldIndex() > tupleTypeSymbol->getResolvedTypes().size())
+      ctx->getFieldIndex() > static_cast<int>(tupleTypeSymbol->getResolvedTypes().size()))
     throw SizeError(ctx->getLineNumber(), "Invalid tuple index");
   return {};
 }
